move cuda code in plugin
