+++ RecoLocalCalo/EcalRecAlgos/BuildFile.xml
@@ -1,6 +1,4 @@
 <use name="clhep"/>
-<use name="cuda"/>
-<use name="CUDADataFormats/EcalRecHitSoA"/>
 <use name="DataFormats/EcalRecHit"/>
 <use name="DataFormats/EcalDigi"/>
 <use name="FWCore/MessageLogger"/>
