+++ RecoLocalCalo/EcalRecProducers/plugins/AmplitudeComputationCommonKernels.cu
