+++ RecoLocalCalo/EcalRecProducers/plugins/AmplitudeComputationCommonKernels.h
