+++ RecoLocalCalo/EcalRecProducers/plugins/AmplitudeComputationKernels.cu
