+++ RecoLocalCalo/EcalRecProducers/plugins/AmplitudeComputationKernels.h
