+++ RecoLocalCalo/EcalRecProducers/plugins/BuildFile.xml
@@ -16,6 +16,6 @@
 <use name="CondFormats/DataRecord"/>
 <use name="RecoLocalCalo/EcalRecAlgos"/>
 <use name="SimCalorimetry/EcalSimAlgos"/>
-<library file="*.cc" name="RecoLocalCaloEcalRecProducersPlugins">
+<library file="*.cc *.cu" name="RecoLocalCaloEcalRecProducersPlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
