+++ RecoLocalCalo/EcalRecProducers/plugins/EcalRecHitBuilderKernels.cu
