+++ RecoLocalCalo/EcalRecProducers/plugins/EcalRecHitBuilderKernels.h
