+++ RecoLocalCalo/EcalRecProducers/plugins/KernelHelpers.cu
