+++ RecoLocalCalo/EcalRecProducers/plugins/KernelHelpers.h
