+++ RecoLocalCalo/EcalRecProducers/plugins/TimeComputationKernels.cu
