+++ RecoLocalCalo/EcalRecProducers/plugins/TimeComputationKernels.h
