+++ RecoLocalCalo/EcalRecProducers/plugins/inplace_fnnls.h
